# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/catalog/ebooks": {
      "get": {
        "operationId": "CatalogController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Titre, auteur, catégorie ou mot-clé (insensible aux accents)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Slug de catégorie",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "title",
                "author"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste paginée des e-books publiés, avec recherche, filtre et tri",
        "tags": [
          "Catalogue"
        ]
      }
    },
...
    "/v1/catalog/ebooks/{slug}": {
      "get": {
        "operationId": "CatalogController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fiche d'un e-book publié (ou possédé, même dépublié)",
        "tags": [
          "Catalogue"
        ]
      }
    },
...
    "/v1/admin/ebooks": {
      "get": {
        "operationId": "AdminEbooksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin — E-books"
        ]
      },
      "post": {
        "operationId": "AdminEbooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEbookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un brouillon",
        "tags": [
          "Admin — E-books"
        ]
      }
    },
    "/v1/admin/ebooks/{id}": {
      "get": {
        "operationId": "AdminEbooksController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin — E-books"
        ]
      },
      "patch": {
        "operationId": "AdminEbooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EbookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin — E-books"
        ]
      },
      "delete": {
        "operationId": "AdminEbooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer définitivement un brouillon jamais commandé",
        "tags": [
          "Admin — E-books"
        ]
      }
    },
    "/v1/admin/ebooks/{id}/files/{kind}": {
      "post": {
        "operationId": "AdminEbooksController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envoyer la couverture (JPEG/PNG/WebP) ou le fichier principal (EPUB/PDF)",
        "tags": [
          "Admin — E-books"
        ]
      }
    },
    "/v1/admin/ebooks/{id}/publish": {
      "post": {
        "operationId": "AdminEbooksController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin — E-books"
        ]
      }
    },
    "/v1/admin/ebooks/{id}/unpublish": {
      "post": {
        "operationId": "AdminEbooksController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dépublier : retiré du catalogue, toujours lisible par les acheteurs (règle D12)",
        "tags": [
          "Admin — E-books"
        ]
      }
    },
    "/v1/admin/ebooks/{id}/archive": {
      "post": {
        "operationId": "AdminEbooksController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin — E-books"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Folio",
    "description": "Vente et lecture d'e-books — application mobile et administration. Toutes les routes sont préfixées par /v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEbookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "description": "Nom de catégorie (créée si absente)"
          },
          "priceCents": {
            "type": "object",
            "description": "Prix TTC en centimes ; vide = prix à définir (non publiable)",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "format": {
            "type": "string"
          },
          "pages": {
            "type": "object"
          },
          "downloadAllowed": {
            "type": "boolean"
          },
          "previewBlocks": {
            "type": "number",
            "description": "Nombre de blocs de texte de l'extrait gratuit"
          },
          "featured": {
            "type": "boolean"
          },
          "coverBg": {
            "type": "string"
          },
          "coverFg": {
            "type": "string"
          },
          "coverTint": {
            "type": "string"
          }
        }
      },
      "EbookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "description": "Nom de catégorie (créée si absente)"
          },
          "priceCents": {
            "type": "object",
            "description": "Prix TTC en centimes ; vide = prix à définir (non publiable)",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "format": {
            "type": "string"
          },
          "pages": {
            "type": "object"
          },
          "downloadAllowed": {
            "type": "boolean"
          },
          "previewBlocks": {
            "type": "number",
            "description": "Nombre de blocs de texte de l'extrait gratuit"
          },
          "featured": {
            "type": "boolean"
          },
          "coverBg": {
            "type": "string"
          },
          "coverFg": {
            "type": "string"
          },
          "coverTint": {
            "type": "string"
          }
        }
      },
...
    }
  }
}